requirements.txt
insightface>=0.1.5
opencv-python>=4.5.1.48
numpy>=1.20.2
pillow>=8.1.2
matplotlib>=3.4.1
python-dateutil>=2.8.1
requests>=2.25.1